refactor: Fix commands module typo.
