+++ .gitignore
@@ -1,3 +1,3 @@
 /target
 /tmp
-*.local.json
+*.local.*
