+++ crates/oneiros/src/commands/doctor.rs
